ChatBot/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
mysql-connector-python==8.1.0
python-dotenv==1.0.0
openai==1.99.7
psycopg2-binary==2.9.7
tqdm==4.67.1
python-multipart==0.0.20
psycopg2-binary==2.9.11
yt_dlp==2025.8.11
pydub==0.25.1
supabase==2.18.0
fdb==2.0.4
firebirdsql==1.4.1
passlib==1.7.4